# mytwip/docs/openagenda/openapi.json
{"openapi":"3.1.0","info":{"title":"MyTwip OpenAgenda API","description":"API for parsing OpenAgenda agendas and avents","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Current API Status","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/agendas/search/{search_term}":{"get":{"summary":"Agendas Search","description":"Search for OpenAgenda agendas by search term.","operationId":"agendas_search_agendas_search__search_term__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"search_term","in":"path","required":true,"schema":{"type":"string","title":"Search Term"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agendas/by_slug/{agenda_slug}":{"get":{"summary":"Agendas By Slug","description":"Search for OpenAgenda agendas by slug","operationId":"agendas_by_slug_agendas_by_slug__agenda_slug__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"agenda_slug","in":"path","required":true,"schema":{"type":"string","title":"Agenda Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agendas/details/{agenda_uid}":{"get":{"summary":"Agendas Details","description":"Search for OpenAgenda agendas by uid","operationId":"agendas_details_agendas_details__agenda_uid__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"agenda_uid","in":"path","required":true,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Agenda Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events/by_agenda_uid/{agenda_uid}":{"get":{"summary":"Events By Agenda Uid","description":"Search for OpenAgenda events by agenda uid","operationId":"events_by_agenda_uid_events_by_agenda_uid__agenda_uid__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"agenda_uid","in":"path","required":true,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Agenda Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cache_query_cleanup":{"get":{"summary":"Cache Query Cleanup","description":"Cleanup or reset query cache","operationId":"cache_query_cleanup_cache_query_cleanup_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"force","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Force"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}